Extracted media player into a service. Background playback is now possible. Main Activity shows implementation of the MediaPlayerService.
