+++ app/src/main/java/open_sound_stream/ossapp/MediaPlayerService.java
@@ -1,0 +1,112 @@
+package open_sound_stream.ossapp;
+
+import android.app.Service;
+import android.content.Intent;
+import android.os.Binder;
+import android.os.Build;
+import android.os.IBinder;
+import android.widget.Button;
+import android.widget.SeekBar;
+
+import androidx.annotation.RequiresApi;
+
+public class MediaPlayerService extends Service {
+
+    private final IBinder mBinder = new LocalBinder();
+
+    public SeekBar mSeekbarAudio;
+    public PlayerAdapter mPlayerAdapter;
+    public boolean mUserIsSeeking;
+
+    public class LocalBinder extends Binder {
+        MediaPlayerService getService() {
+            return MediaPlayerService.this;
+        }
+    }
+
+    @Override
+    public IBinder onBind(Intent intent) {
+        return mBinder;
+    }
+
+    public void initializeUI(Button mPlayPauseButton, Button mPrevButton, Button mNextButton, SeekBar seekBar) {
+        mSeekbarAudio = seekBar;
+
+        mPlayPauseButton.setOnClickListener(
+                view -> {
+                    if (mPlayerAdapter.isPlaying()) {
+                        mPlayerAdapter.pause();
+                    }
+                    else {
+                        mPlayerAdapter.play();
+                    }
+                });
+
+        mPrevButton.setOnClickListener(
+                view -> {
+                    if (mPlayerAdapter.getCurrentPlaybackPosition() <= 5000)
+                    {
+                        //to previous track
+                    } else {
+                        mPlayerAdapter.seekTo(0);
+                    }
+                });
+        initializeSeekbar();
+        initializePlaybackController();
+    }
+
+    public void initializePlaybackController() {
+        MediaPlayerHolder mMediaPlayerHolder = new MediaPlayerHolder(this);
+        mMediaPlayerHolder.setPlaybackInfoListener(new MediaPlayerService.PlaybackListener());
+        mPlayerAdapter = mMediaPlayerHolder;
+    }
+
+    public void initializeSeekbar() {
+        mSeekbarAudio.setOnSeekBarChangeListener(
+                new SeekBar.OnSeekBarChangeListener() {
+                    int userSelectedPosition = 0;
+
+                    @Override
+                    public void onStartTrackingTouch(SeekBar seekBar) {
+                        mUserIsSeeking = true;
+                    }
+
+                    @Override
+                    public void onProgressChanged(SeekBar seekBar, int progress, boolean fromUser) {
+                        if (fromUser) {
+                            userSelectedPosition = progress;
+                        }
+                    }
+
+                    @Override
+                    public void onStopTrackingTouch(SeekBar seekBar) {
+                        mUserIsSeeking = false;
+                        mPlayerAdapter.seekTo(userSelectedPosition);
+                    }
+                });
+    }
+
+    public class PlaybackListener extends PlaybackInfoListener {
+
+        @Override
+        public void onDurationChanged(int duration) {
+            mSeekbarAudio.setMax(duration);
+        }
+
+        @RequiresApi(api = Build.VERSION_CODES.N)
+        @Override
+        public void onPositionChanged(int position) {
+            if (!mUserIsSeeking) {
+                mSeekbarAudio.setProgress(position, true);
+            }
+        }
+
+        @Override
+        public void onStateChanged(@State int state) {
+        }
+
+        @Override
+        public void onPlaybackCompleted() {
+        }
+    }
+}
